+++ lib/mixins/movement.dart
@@ -168,15 +168,7 @@ mixin Movement on GameComponent {
   /// Move Player to direction by radAngle
   void moveFromAngle(double angle, {double? speed}) {
     _lastSpeed = speed ?? this.speed;
-    var x = cos(angle) * _lastSpeed;
-    var y = sin(angle) * _lastSpeed;
-    if (x.abs() < 0.01) {
-      x = 0;
-    }
-    if (y.abs() < 0.01) {
-      y = 0;
-    }
-    velocity = Vector2(x, y);
+    velocity = BonfireUtil.vector2ByAngle(angle, intencity: _lastSpeed);
   }
 
   void stopMove({bool forceIdle = false, bool isX = true, bool isY = true}) {