+++ lib/util/bonfire_util.dart
@@ -104,4 +104,16 @@ class BonfireUtil {
   ) {
     return movePointByAngle(point, speed, angle) - point;
   }
+
+  static Vector2 vector2ByAngle(double angle, {double intencity = 1}) {
+    var x = cos(angle) * intencity;
+    var y = sin(angle) * intencity;
+    if (x.abs() < 0.01) {
+      x = 0;
+    }
+    if (y.abs() < 0.01) {
+      y = 0;
+    }
+    return Vector2(x, y);
+  }
 }