+++ lib/util/extensions/movement_extensions.dart
@@ -1,8 +1,4 @@
-import 'dart:ui';
-
-import 'package:bonfire/base/game_component.dart';
-import 'package:bonfire/mixins/movement.dart';
-import 'package:bonfire/util/extensions/game_component_extensions.dart';
+import 'package:bonfire/bonfire.dart';
 
 extension MovementExtensions on Movement {
   /// This method move this component to target
@@ -20,7 +16,6 @@ extension MovementExtensions on Movement {
 
     if (rectCollision.overlaps(rectPlayerCollision)) {
       close?.call();
-      moveFromAngle(radAngle);
       stopMove();
       return false;
     }