+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# NEXT
+- `Follower` mixin improvements.
+- Adds `startFollowPlayer` param in `CameraConfig`.
+- `Vision` improvements.
+
 # 3.0.0-beta.9
 - Adds `getZoomFromMaxVisibleTile`
 - `moveFromAngle` improvements.